LayoutTests/http/tests/security/window-named-valueOf-expected.txt

CONSOLE MESSAGE: line 1: Threw exception: TypeError: Illegal constructor
This passes if it doesn't alert the contents of innocent-victim.  
